pyproject.toml
[project]
name = "love-guru-agent"
version = "0.1.0"
description = "Enterprise FastAPI architecture for a love coaching AI assistant."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "psycopg[binary]>=3.2.3",
    "pgvector>=0.3.6",
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.18",
    "pyyaml>=6.0.2",
    "langchain>=0.3.7",
    "langchain-core>=0.3.19",
    "langgraph>=0.2.43",
    "langchain-openai>=0.2.8",
    "dashscope>=1.20.14",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
